+++ i18n/src/test/java/net/time4j/format/expert/MiscellaneousTest.java
@@ -54,6 +54,17 @@ import static org.junit.Assert.fail;
 public class MiscellaneousTest {
 
     @Test
+    public void parseMeridiem() throws ParseException {
+        ChronoFormatter<PlainTime> f = ChronoFormatter.ofTimePattern("hh:mm a", PatternType.CLDR, new Locale("es"));
+        assertThat(f.parse("04:30 am"), is(PlainTime.of(4, 30)));
+        assertThat(f.parse("04:30 pm"), is(PlainTime.of(16, 30)));
+        assertThat(f.parse("04:30 AM"), is(PlainTime.of(4, 30)));
+        assertThat(f.parse("04:30 PM"), is(PlainTime.of(16, 30)));
+        assertThat(f.parse("04:30 a. m."), is(PlainTime.of(4, 30)));
+        assertThat(f.parse("04:30 p. m."), is(PlainTime.of(16, 30)));
+    }
+
+    @Test
     public void arabicNuExtensionTest() throws ParseException {
         String input = "الاثنين 24 أبريل 2017 - 15:00";
 
@@ -149,7 +160,7 @@ public class MiscellaneousTest {
     @Test(expected=IllegalArgumentException.class)
     public void printWithGenericPattern() {
         ChronoFormatter<PlainDate> formatter =
-            ChronoFormatter.ofDatePattern("EEEE, yyyy-MM-dd", PatternType.NON_ISO_DATE, Locale.GERMAN);
+            ChronoFormatter.ofDatePattern("EEEE, yyyy-MM-dd", PatternType.CLDR_DATE, Locale.GERMAN);
         formatter.format(PlainDate.of(2015, 7, 17));
     }
 
@@ -855,7 +866,7 @@ public class MiscellaneousTest {
                     embedded)
                 .build()
                 .withDefault(PlainDate.YEAR, 2016)
-                .withDefault(PlainTime.ISO_HOUR, 0);
+                .withDefault(PlainTime.HOUR_FROM_0_TO_24, 0);
         PlainTimestamp tsp = f.parse("02-29");
         assertThat(tsp, is(PlainTimestamp.of(2016, 2, 29, 0, 0)));
     }